requirements.txt
Authlib==1.6.6
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
Flask==3.1.2
Flask-JWT-Extended==4.7.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.3
mongoengine==0.29.1
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
stripe==14.1.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
